sql query from range of column values
A helper function to construct a sql query for a column that can take a
range of ordinal values. For example an income column with many subsets
of income ranges (i.e., 15,000 - 25,000 or 100,000 - 150,000, etc.).
Store the complete range of values in an ordered list and pass the
beginning and ending values to construct a query string of each
encompassed value separated by a logical OR operator.